String array in SquareTable Method
The Octets property in SquareTable method is now a string array storing the "octets" values in a string form at each indice of the Octets array.

I am using the term octets to refer to the set of  squared values of integers between 0 and 1000 where the last 3 digits  are the same.